refactors bootstrap_geography. Adds bootstrap for HI. Adds shapefile for HI. Roll 0.4.2
