+++ geography/bin/hi/clipped_hi.cpg
