+++ geography/bin/pa/clipped_pa.cpg
@@ -1,0 +1,1 @@
+UTF-8
