+++ geography/bin/pa/clipped_pa.shp
