+++ geography/management/commands/bootstrap/download/__init__.py
@@ -1,0 +1,4 @@
+from ._county import DownloadCounty
+from ._district import DownloadDistrict
+from ._township import DownloadTownship
+from ._state import DownloadState
