+++ geography/management/commands/bootstrap/fixtures/__init__.py
@@ -1,0 +1,4 @@
+from ._county import CountyFixtures
+from ._district import DistrictFixtures
+from ._nation import NationFixtures
+from ._state import StateFixtures
