+++ geography/management/commands/bootstrap/shapes/__init__.py
@@ -1,0 +1,3 @@
+from ._county import CountyShape
+from ._district import DistrictShape
+from ._township import TownshipShape
