+++ geography/management/commands/bootstrap_geom_hi.py
@@ -1,0 +1,85 @@
+import os
+
+from tqdm import tqdm
+
+import geojson
+import shapefile
+from django.core.management.base import BaseCommand
+from geography.models import Division, DivisionLevel, Geometry
+from geography.utils.lookups import county_lookup
+
+from .bootstrap._arguments import ArgumentsMethods
+from .bootstrap._attributes import Attributes
+from .bootstrap._toposimplify import Toposimplify
+
+SHP_SLUG = 'clipped_hi'
+STATE_LEVEL = DivisionLevel.STATE
+
+
+class Command(Toposimplify, ArgumentsMethods, Attributes, BaseCommand):
+
+    def get_county_shp(self, fips):
+        cmd_path = os.path.dirname(os.path.realpath(__file__))
+        SHAPEFILE_PATH = os.path.join(cmd_path, '../../bin/hi')
+
+        shape = shapefile.Reader(os.path.join(
+            SHAPEFILE_PATH,
+            '{}.shp'.format(SHP_SLUG)
+        ))
+        fields = shape.fields[1:]
+        field_names = [f[0] for f in fields]
+
+        county_records = [
+            shp for shp in shape.shapeRecords()
+            if dict(zip(field_names, shp.record))['STATEFP'] == fips or
+            (
+                fips == '00' and
+                int(dict(zip(field_names, shp.record))['STATEFP']) <= 56
+            )
+        ]
+        features = []
+        for shp in county_records:
+            rec = dict(zip(field_names, shp.record))
+            geometry = shp.shape.__geo_interface__
+            geodata = {
+                'type': 'Feature',
+                'geometry': geometry,
+                'properties': {
+                    'state': rec['STATEFP'],
+                    'county': rec['COUNTYFP'],
+                    'name': county_lookup[rec['STATEFP']].get(
+                        rec['COUNTYFP'], rec['NAME']
+                    )
+                }
+            }
+            features.append(geodata)
+        threshold = self.THRESHOLDS['nation'] if fips == '00' else \
+            self.THRESHOLDS['county']
+        return self.toposimplify(
+            geojson.FeatureCollection(features),
+            threshold
+        )
+
+    def handle(self, *args, **options):
+        self.set_attributes()
+        self.parse_arguments(options)
+
+        state = Division.objects.get(
+            code=15,
+            level__name=DivisionLevel.STATE
+        )
+
+        geojson, created = Geometry.objects.update_or_create(
+            division=state,
+            subdivision_level=self.COUNTY_LEVEL,
+            simplification=0.075,
+            source='https://www2.census.gov/geo/tiger/GENZ2016/shp/cb_2016_us_state_500k.zip',
+            series='2016',
+            defaults={
+                'topojson': self.get_county_shp('15'),
+            },
+        )
+
+        self.stdout.write(
+            self.style.SUCCESS('All done! 🏁')
+        )
